Remove more standalone vars, typedefs, includes from Config.h.
